get data entry

+++ src/utils.rs
@@ -5,8 +5,8 @@ pub fn hash(input: &str) -> String {
     format!("{:x}", md5::compute(input))
 }
 
-pub fn unixtime_to_datetime(timestamp: u64) -> String {
-    let secs = UNIX_EPOCH + Duration::from_secs(timestamp);
+pub fn unixtime_to_datetime(timestamp: u32) -> String {
+    let secs = UNIX_EPOCH + Duration::from_secs(u64::from(timestamp));
 
     DateTime::<Utc>::from(secs).format("%Y/%m/%d").to_string()
 }
